added gui and executable spec

+++ gui_main.pyw
@@ -1,7 +1,7 @@
 import os, sys
 
 from tkinter import *
-from tkinter import filedialog, ttk
+from tkinter import filedialog, ttk, messagebox
 from tkinter.scrolledtext import ScrolledText
 
 from huffman_encoder import (
@@ -9,18 +9,31 @@ from huffman_encoder import (
     decode_bin
 )
 
+# configure root directory path relative to this file
+THIS_FOLDER_G = ""
+if getattr(sys, "frozen", False):
+    # frozen
+    THIS_FOLDER_G = os.path.dirname(sys.executable)
+else:
+    # unfrozen
+    THIS_FOLDER_G = os.path.dirname(os.path.realpath(__file__))
+
+
 class CompressFrame(Frame):
-    def __init__(self):
-        super().__init__()
+
+    def __init__(self, parent):
+        super().__init__(parent)
         self.render_select_file_section()
 
     def render_select_file_section(self):
-        self.title_label = Label(self, bg='#ffffff', text='COMPRESSION', anchor=CENTER, relief=FLAT)
-        self.title_label.grid(row=0, column=0, columnspan=2, padx=12, pady=10, sticky=W+E+N+S)
+        self.title_label = Label(self, bg='#ffffff',
+                                 text='COMPRESSION ~ Select a File or Enter its Path',
+                                 anchor=W, relief=FLAT)
+        self.title_label.grid(row=0, column=0, columnspan=2, padx=10, pady=(14, 7), sticky=W+E+N+S)
 
 
-        self.input_file_entry = Entry(self, width=50, relief=FLAT, bg='#eee', fg='#333',
-                                      selectbackground='#ffffff', selectforeground='#333',
+        self.input_file_entry = Entry(self, width=54, relief=FLAT, bg='#eee', fg='#333',
+                                      selectbackground='#feffc6', selectforeground='#333',
                                       validate='key', exportselection=False)
         self.input_file_entry.grid(row=1, column=0, padx=(12, 6), pady=6, ipadx=12, ipady=6)
         self.input_file_entry.delete(0, END)
@@ -41,7 +54,7 @@ class CompressFrame(Frame):
 
         
         self.status_label = Label(self, bg='#ffffff', text='---', anchor=W, relief=FLAT)
-        self.status_label.grid(row=3, column=0, columnspan=2, padx=12, pady=6, ipadx=12, ipady=0, sticky=W+E+N+S)
+        self.status_label.grid(row=3, column=0, columnspan=2, padx=10, pady=6, ipadx=12, ipady=0, sticky=W+E+N+S)
         
 
         self.config(bg='#ffffff')
@@ -57,7 +70,8 @@ class CompressFrame(Frame):
         if os.path.isfile(P):
             self.compress_file_btn.config(state=NORMAL, relief=FLAT, bg='#1fc186')
         else:
-            self.compress_file_btn.config(state=DISABLED, relief=GROOVE, bg='#ffffff')
+            self.compress_file_btn.config(state=DISABLED, relief=FLAT, bg='#ddd')
+        self.status_label.config(text='---')
         return True
     
     def compress_btn_callback(self):
@@ -81,28 +95,36 @@ class CompressFrame(Frame):
                           .asksaveasfilename(initialdir=dir_name,
                                              title="Save Compressed File",
                                              filetypes=(("Compressed File","*.hc1"),("all files","*.*")))
-
+        
+        isfilepath_selected = True if output_filepath != '' else False
+        
         output_filepath = output_filepath if output_filepath.split('.')[-1] == ext else output_filepath + '.' + ext
         output_filepath = output_filepath if output_filepath.split('.')[-1] == 'hc1' else output_filepath + '.hc1'
         
-        with open(output_filepath, 'wb') as output_file:
-            output_file.write(compressed_data)
+        if isfilepath_selected:
+            self.status_label.config(text='Compressing ...')
 
-        self.status_label.config(text='SUCCESS: File Compressed Successfully!')
+            with open(output_filepath, 'wb') as output_file:
+                output_file.write(compressed_data)
+
+            self.status_label.config(text='SUCCESS: File Compressed Successfully!')
 
 
 class DecompressFrame(Frame):
-    def __init__(self):
-        super().__init__()
+
+    def __init__(self, parent):
+        super().__init__(parent)
         self.render_select_file_section()
 
     def render_select_file_section(self):
-        self.title_label = Label(self, bg='#ffffff', text='DECOMPRESSION', anchor=CENTER, relief=FLAT)
-        self.title_label.grid(row=0, column=0, columnspan=2, padx=12, pady=10, sticky=W+E+N+S)
+        self.title_label = Label(self, bg='#ffffff',
+                                 text='DECOMPRESSION ~ Select a File or Enter its Path',
+                                 anchor=W, relief=FLAT)
+        self.title_label.grid(row=0, column=0, columnspan=2, padx=10, pady=(14, 7), sticky=W+E+N+S)
 
         
-        self.input_file_entry = Entry(self, width=50, relief=FLAT, bg='#eee', fg='#333',
-                                      selectbackground='#ffffff', selectforeground='#333',
+        self.input_file_entry = Entry(self, width=54, relief=FLAT, bg='#eee', fg='#333',
+                                      selectbackground='#feffc6', selectforeground='#333',
                                       validate='key', exportselection=False)
         self.input_file_entry.grid(row=1, column=0, padx=(12, 6), pady=6, ipadx=12, ipady=6)
         self.input_file_entry.delete(0, END)
@@ -123,7 +145,7 @@ class DecompressFrame(Frame):
 
         
         self.status_label = Label(self, bg='#ffffff', text='---', anchor=W, relief=FLAT)
-        self.status_label.grid(row=3, column=0, columnspan=2, padx=12, pady=6, ipadx=12, ipady=0, sticky=W+E+N+S)
+        self.status_label.grid(row=3, column=0, columnspan=2, padx=10, pady=6, ipadx=12, ipady=0, sticky=W+E+N+S)
         
 
         self.config(bg='#ffffff')
@@ -139,7 +161,8 @@ class DecompressFrame(Frame):
         if os.path.isfile(P):
             self.decompress_file_btn.config(state=NORMAL, relief=FLAT, bg='#1fc186')
         else:
-            self.decompress_file_btn.config(state=DISABLED, relief=GROOVE, bg='#ffffff')
+            self.decompress_file_btn.config(state=DISABLED, relief=FLAT, bg='#ddd')
+        self.status_label.config(text='---')
         return True
     
     def decompress_btn_callback(self):
@@ -164,23 +187,64 @@ class DecompressFrame(Frame):
                                              title="Save Decompressed File",
                                              filetypes=(("." + ext,"." + ext),("all files","*.*")))
 
+        isfilepath_selected = True if output_filepath != '' else False
+        
         output_filepath = output_filepath if output_filepath.split('.')[-1] == ext else output_filepath + '.' + ext
         # output_filepath = output_filepath if output_filepath.split('.')[-1] == 'hc1' else output_filepath + '.hc1'
         
-        with open(output_filepath, 'wb') as output_file:
-            output_file.write(decompressed_data)
-        
-        self.status_label.config(text='SUCCESS: File Decompressed Successfully!')
+        if isfilepath_selected:
+            self.status_label.config(text='Decompressing ...')
+
+            with open(output_filepath, 'wb') as output_file:
+                output_file.write(decompressed_data)
+            
+            self.status_label.config(text='SUCCESS: File Decompressed Successfully!')
+
+
+class MenuBar(Menu):
+
+    def __init__(self, parent):
+        super().__init__(parent)
+
+        self.configure(
+            bg="#ddd",
+            relief=FLAT
+        )
+        self.add_command(
+            label="About",
+            command=self.show_about
+        )
+        self.add_command(
+            label="Quit!",
+            command=parent.quit
+        )
+
+    def show_about(self):
+        messagebox.showinfo(
+        "About",
+        """A File Compression / Decompression Tool based on Huffman Encoding
+
+Source Code:
+https://github.com/ahmednooor/ridiculously_slow_huffman_encoder""")
 
 
 def main():
     root = Tk()
-    root.title('Compression Tool ~ Huffman Encoder Decoder')
+    root.title('File Compression Tool ~ Huffman Encoder Decoder')
     root.config(bg='#efefef')
+
+    try:
+        root.iconbitmap(THIS_FOLDER_G + '/assets/icon.ico')
+    except:
+        pass
+
+    menu_bar = MenuBar(root)
+    root.configure(menu=menu_bar)
+
     tabs = ttk.Notebook(root)
 
-    compress_tab = CompressFrame()
-    decompress_tab = DecompressFrame()
+    compress_tab = CompressFrame(tabs)
+    decompress_tab = DecompressFrame(tabs)
     
     tabs.add(compress_tab, text='Compress')
     tabs.add(decompress_tab, text='Decompress')
